progressing

+++ src/app/table/table.component.ts
@@ -18,29 +18,57 @@ export class TableComponent implements OnInit {
   total = 0;
   productPerPageArr = [5, 10, 15, 20];
   productPerPage: number = 5;
-  loginEditReactiveForm: FormGroup;
+  loginReactiveForm: FormGroup;
+
   searchword = '';
 
   constructor(private _storageService: StorageService) {}
 
-  loginReactiveForm = new FormGroup({
-    name: new FormControl('', [
-      Validators.required,
-      Validators.pattern('[a-zA-Z]*'),
-    ]),
-    description: new FormControl('', [
-      Validators.required,
-      Validators.pattern('[a-zA-Z]*'),
-    ]),
-    image: new FormControl('', [Validators.required]),
-    rating: new FormControl('', [Validators.required]),
-    select: new FormControl('', [Validators.required]),
-    radioinput: new FormControl('', [Validators.required]),
-    price: new FormControl('', [Validators.required]),
-    quantity: new FormControl('', [Validators.required]),
-  });
+  makeForm(id?:any) {
+    if(!this.isEditForm) {
+      this.loginReactiveForm = new FormGroup({
+        name: new FormControl('', [
+          Validators.required,
+          Validators.pattern('[a-zA-Z]*'),
+        ]),
+        description: new FormControl('', [
+          Validators.required,
+          Validators.pattern('[a-zA-Z]*'),
+        ]),
+        image: new FormControl('', [Validators.required]),
+        rating: new FormControl('', [Validators.required]),
+        inventory_status: new FormControl('', [Validators.required]),
+        category: new FormControl('', [Validators.required]),
+        price: new FormControl('', [Validators.required]),
+        quantity: new FormControl('', [Validators.required]),
+      });
+    } else {
+      this.editData = this.showData.find((ele) => ele.code == id);
+      this.loginReactiveForm = new FormGroup({
+        name: new FormControl(this.editData['name'], [
+          Validators.required,
+          Validators.pattern('[a-zA-Z]*'),
+        ]),
+        description: new FormControl(this.editData['description'], [
+          Validators.required,
+          Validators.pattern('[a-zA-Z]*'),
+        ]),
+        image: new FormControl(this.editData['image'], [Validators.required]),
+        rating: new FormControl(this.editData['rating'], [Validators.required]),
+        inventory_status: new FormControl(this.editData['inventory_status'], [Validators.required]),
+        category: new FormControl(this.editData['category'], [
+          Validators.required,
+        ]),
+        price: new FormControl(this.editData['price'], [Validators.required]),
+        quantity: new FormControl(this.editData['quantity'], [
+          Validators.required,
+        ]),
+      });
+    }
+  }
 
   ngOnInit(): void {
+    this.makeForm();
     this.loadData();
   }
 
@@ -49,32 +77,36 @@ export class TableComponent implements OnInit {
       this.showData = this._storageService.getItemsFromLocalStorage();
       this.total = this.showData.length;
     } else {
+      console.log("this searchword", this.searchword);
+      
       this.showData = this.showData.filter((ele) =>
-        ele.name.includes(this.searchword)
+        ele.code.includes(this.searchword) || 
+        ele.name.includes(this.searchword) ||
+        ele.rating == +this.searchword ||
+        ele.inventory_status.includes(this.searchword.toUpperCase()) ||
+        ele.category.includes(this.searchword) 
       );
       this.total = this.showData.length;
     }
   }
-  submitEditData() {
-    console.log(this.loginEditReactiveForm.value);
-    const index = this.showData.findIndex(
-      (ele) => ele.code == this.editData.code
-    );
-    this.showData[index] = this.loginEditReactiveForm.value;
-    this.isEditForm = !this.isEditForm;
-    this._storageService.editItemInLocalStorage(
-      this.loginEditReactiveForm.value,
-      this.editData.code
-    );
-  }
-  submitData(myForm: any) {
+  
+  submitData() {
     console.log('hello in submit');
-    // this.showData.push(this.loginReactiveForm.value);
-    this._storageService.setItemInLocalStorage(this.loginReactiveForm.value);
+    if(!this.isEditForm) {
+      this._storageService.setItemInLocalStorage(this.loginReactiveForm.value);
     this.isAddForm = !this.isAddForm;
     this.total = this.total + 1;
-    myForm.value = '';
     this.loadData();
+    } else {
+      const index = this.showData.findIndex(
+        (ele) => ele.code == this.editData.code
+      );
+      this.showData[index] = this.loginReactiveForm.value;
+      this._storageService.editItemInLocalStorage(
+        this.loginReactiveForm.value,
+        this.editData.code
+      );
+    }
     this.loginReactiveForm.reset();
   }
 
@@ -84,41 +116,20 @@ export class TableComponent implements OnInit {
     this._storageService.editItemInLocalStorage(this.showData[index], id);
   }
   deleteRow(id: string) {
-    this.showData = this.showData.filter((ele) => ele.code != id);
-    this.total = this.showData.length;
-    this._storageService.deleteItemFromLocalStorage(id);
-  }
+      if(confirm('Are you sure you want to delete')) {
+        this.showData = this.showData.filter((ele) => ele.code != id);
+        this.total = this.showData.length;
+        this._storageService.deleteItemFromLocalStorage(id);  
+      }
 
-  openEditForm(id: number) {
-    this.isEditForm = !this.isEditForm;
-    this.editData = this.showData.find((ele) => ele.code == id);
-    this.loginEditReactiveForm = new FormGroup({
-      name: new FormControl(this.editData['name'], [
-        Validators.required,
-        Validators.pattern('[a-zA-Z]*'),
-      ]),
-      description: new FormControl(this.editData['description'], [
-        Validators.required,
-        Validators.pattern('[a-zA-Z]*'),
-      ]),
-      image: new FormControl(this.editData['image'], [Validators.required]),
-      rating: new FormControl(this.editData['rating'], [Validators.required]),
-      select: new FormControl(this.editData['select'], [Validators.required]),
-      radioinput: new FormControl(this.editData['radioinput'], [
-        Validators.required,
-      ]),
-      price: new FormControl(this.editData['price'], [Validators.required]),
-      quantity: new FormControl(this.editData['quantity'], [
-        Validators.required,
-      ]),
-    });
-    console.log(this.editData);
   }
-
+  
   deleteSelected() {
+    if(confirm('Are you sure you want to delete')) {
     this.showData = this.showData.filter((ele) => ele.isCheck == false);
     this._storageService.deleteSelectedInLocStorage();
     this.total = this.showData.length;
+    }
   }
 
   selectAll(data: any) {
@@ -131,9 +142,16 @@ export class TableComponent implements OnInit {
     this.isAddForm = true;
   }
 
+  openEditForm(id: any) {
+    this.isEditForm = true;
+    this.makeForm(id);
+    console.log(this.editData);
+  }
+
   closeForm() {
     this.isAddForm = false;
     this.isEditForm = false;
+    this.loginReactiveForm.reset();
   }
 
   sortList(sortVal: string) {
@@ -143,8 +161,9 @@ export class TableComponent implements OnInit {
     this.arrow = this.arrow == 'up' ? 'down' : 'up';
   }
 
-  searchWord(word: string) {
+  searchInput(word: string) {
     this.searchword = word;
+    
     this.loadData();
   }
 
@@ -162,8 +181,6 @@ export class TableComponent implements OnInit {
       quoteStrings: '"',
       decimalseparator: '.',
       showLabels: true,
-      showTitle: true,
-      title: 'product list',
       useBom: true,
       noDownload: false,
       headers: ['Name', 'Description', 'Image', 'Rating', 'Incventory Status', 'Category', 'Price', 'Quantity', 'ID', 'CheckBox'],
